Plugin preference UI elements
Started working on a PreferenceCategory/Preference that apps can use.
Plugin loading is too slow to be done on every start and UI population.
A persistent cache is needed!

Also:
Added debug flag that prevents disabling of plugin components on
crashes, which can be annoying for app and plugin developers.