In-between upload
TODO: Everything shared between InstanceManager and pluginFactoryManager,
including PluginInfo/PluginFactoryInfo and so on, should be made generic
and possibly an abstract class, so that you an load "items" from APKs.
The subclasses shall create the items/instanced, either when asked to
(startPlugin) or when loaded (PluginFactory).
Possibly move away from the word "plugin" and "item".
One thing to solve is the parameters to the manager(s) - plugins need
the KClass<Plugin>+action but PluginFactory only needs action. Or can
they be generic enough that the specifics are wrapped in a generic
handler that you et in a subclass, or something? The factory for the
manager instances?