Big in-between upload
Renamed PluginInfo to InstanceInfo and took away all instantiation inside
InstanceManagerImpl. Instead, the idea is that a plugin class is not
instantiated until it is started.
PluginInfo class anme has been exposed outwards as a class to get all
metadata and control the lifecycle of the Plugin subclass. The client
has to call PluginInfo.start() to instantiate the plugin. Once it's
instantiated it can possibly be re-used, but it's probably safer to
let it fall to the GC after PluginIndo.stop().