Back to KClass
Go back to KClass from String, instead we preprocess the mapping
from String to KClass in the annotation processor.

PluginInfo now wraps one instance of a Plugin. It has a reference
to the InstanceInfo (of which there is only one) but there will be
one PluginInstance for each time getPlugin/getPligunList is called.
teh client will then call start() on these instances when needed.