Advanced PluginInfo handling
Various components now manipulate PluginInfo to allow the client app
to pass along extra parameters when opening plugin settings and in
other places. This should make stuff like sessions easier without
having to rely on session switching (which was crude and could lead
to some nasty bugs in the future).