Add PluginInfo hooks
CLient can add hooks around code that will create PluginInfo instances, and they
will get a callback when the PluginInfo is created so that they can for example
add data to it that will follow its lifecycle.

ALso refactor the public part of the library to match the implementation part a
bit better.